добавленна пагинация , на главную и поиск,
