+++ src/index/index.js
@@ -1,7 +1,6 @@
 import '../sass/main.scss';
 import '../js/trendFilms';
-import '../js/pagination.vova';
-import request from '../js/apiRequest.js';
+import '../js/pagination';
 import '../js/getGenres';
 import '../js/movieSearch';
 import '../js/myLibrary';
