+++ src/js/movieSearch.js
@@ -3,17 +3,50 @@ import request from '../js/apiRequest';
 import debounce from 'lodash.debounce';
 import createGallery from './trendFilms';
 import Pagination from 'tui-pagination';
-// import '../tools/styles.css';
+import refs from './refs'
 
-import { container, getPaginationOptions } from './pagination.vova';
+import { container, getTotalPages } from './pagination';
 
-let query = '';
-const input = document.querySelector('input');
-const form = document.querySelector('#search-form');
-const galleryList = document.querySelector('.gallery-list'); // замінить клас
+// замінить клас
+
+const pagination = new Pagination(container);
+let query = ''
 
 searchFilms();
 
+async function searchFilms() {
+
+  refs.input.addEventListener(
+    'input',
+    debounce(async () => {
+      query = refs.input.value
+      const data = await request.searchFilms(refs.input.value)
+      refs.galleryList.innerHTML = '';
+
+      createGallery(data);
+      pagination.reset(getTotalPages(data));
+
+    }, 1000)
+  )
+}
+
+pagination.on('beforeMove', async ({ page }) => {
+  refs.galleryList.innerHTML = '';
+
+  refs.input.addEventListener('input', e => {
+    if (e.value !== query) {
+      request.setPage(1)
+    }
+  })
+
+  request.setPage(page)
+  const data = await request.searchFilms(refs.input.value)
+  createGallery(data);
+})
+
+
+
+// ==============================================
 // form.addEventListener('submit', renderImages);
 // function renderImages(e) {
 //   e.preventDefault();
@@ -25,24 +58,3 @@ searchFilms();
 //     form.reset();
 //   });
 // }
-
-function searchFilms() {
-  input.addEventListener(
-    'input',
-    debounce(() => {
-      const inputValue = input.value;
-
-      request.searchFilms(inputValue).then(data => {
-        galleryList.innerHTML = '';
-
-        const pagination = new Pagination(
-          container,
-          getPaginationOptions(data),
-        );
-
-        createGallery(data);
-        pagination.reset();
-      });
-    }, 1000),
-  );
-}
