+++ src/js/refs.js
@@ -1,3 +1,5 @@
 export default {
   gallery: document.querySelector('.gallery-list'),
+  input : document.querySelector('input'),
+  galleryList: document.querySelector('.gallery-list')
 };
